requirements.txt
absl-py==1.0.0
ale-py==0.7.3
argon2-cffi==21.1.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
blobfile==1.2.7
box2d-py==2.3.8
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.0.0
cmake==3.22.0
colorama==0.4.4
crypto==1.4.1
cryptography==36.0.1
cycler==0.11.0
DALL-E==0.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.4.0
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.2
ftfy==6.0.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gym==0.21.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mypy==0.910
mypy-extensions==0.4.3
Naked==0.1.31
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.6
numpy==1.21.2
oauthlib==3.1.1
openai==0.11.0
opencv-python==4.5.5.62
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandas-stubs==1.2.0.39
pandocfilters==1.5.0
parso==0.8.2
passlib==1.7.4
patsy==0.5.2
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.19.1
pure-python-adb==0.3.0.dev0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pycryptodomex==3.11.0
pygame==2.1.0
pyglet==1.5.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sense-hat==2.2.0
setuptools-scm==6.3.2
shellescape==3.8.1
six==1.16.0
statsmodels==0.13.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
xmltodict==0.12.0
zipp==3.6.0
